requirements.txt
requests
pillow
numpy
scikit-learn